Move extractText common to extract-text
